+++ src/main/java/com/yanny/ytech/ModBusSubscriber.java
@@ -1,6 +1,7 @@
 package com.yanny.ytech;
 
 import com.yanny.ytech.configuration.SpearType;
+import com.yanny.ytech.configuration.Utils;
 import com.yanny.ytech.configuration.block_entity.WellPulleyBlockEntity;
 import com.yanny.ytech.configuration.entity.*;
 import com.yanny.ytech.configuration.item.BasketItem;
@@ -11,6 +12,7 @@ import com.yanny.ytech.configuration.renderer.*;
 import com.yanny.ytech.configuration.screen.AqueductFertilizerScreen;
 import com.yanny.ytech.configuration.screen.PrimitiveAlloySmelterScreen;
 import com.yanny.ytech.configuration.screen.PrimitiveSmelterScreen;
+import com.yanny.ytech.configuration.screen.StrainerScreen;
 import com.yanny.ytech.configuration.tooltip.BasketTooltip;
 import com.yanny.ytech.configuration.tooltip.ClientBasketTooltip;
 import com.yanny.ytech.network.irrigation.IrrigationServerNetwork;
@@ -83,6 +85,7 @@ public class ModBusSubscriber {
         event.register(YTechMenuTypes.AQUEDUCT_FERTILIZER.get(), AqueductFertilizerScreen::new);
         event.register(YTechMenuTypes.PRIMITIVE_ALLOY_SMELTER.get(), PrimitiveAlloySmelterScreen::new);
         event.register(YTechMenuTypes.PRIMITIVE_SMELTER.get(), PrimitiveSmelterScreen::new);
+        event.register(YTechMenuTypes.STRAINER.get(), StrainerScreen::new);
     }
 
     @OnlyIn(Dist.CLIENT)
@@ -115,6 +118,7 @@ public class ModBusSubscriber {
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.FIRE_PIT.get(), FirePitRenderer::new);
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.MILLSTONE.get(), MillstoneRenderer::new);
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.POTTERS_WHEEL.get(), PottersWheelRenderer::new);
+        event.registerBlockEntityRenderer(YTechBlockEntityTypes.STRAINER.get(), StrainerRenderer::new);
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.TANNING_RACK.get(), TanningRackRenderer::new);
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.TOOL_RACK.get(), ToolRackRenderer::new);
         event.registerBlockEntityRenderer(YTechBlockEntityTypes.TREE_STUMP.get(), TreeStumpRenderer::new);
@@ -163,6 +167,8 @@ public class ModBusSubscriber {
         for (SpearType spearType : SpearType.values()) {
             event.register(MODEL_IN_HAND_LOCATIONS.get(spearType));
         }
+
+        event.register(ModelResourceLocation.standalone(Utils.modBlockLoc("strainer_net")));
     }
 
     @SubscribeEvent