+++ src/main/java/com/yanny/ytech/configuration/block/AqueductFertilizerBlock.java
@@ -6,6 +6,7 @@ import com.yanny.ytech.registration.YTechBlocks;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.Containers;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
@@ -37,6 +38,17 @@ public class AqueductFertilizerBlock extends AqueductHydratorBlock {
             player.openMenu(Objects.requireNonNull(getMenuProvider(state, level, pos)), pos);
         }
         return InteractionResult.SUCCESS;
+    }
+
+    @Override
+    public void onRemove(@NotNull BlockState state, @NotNull Level level, @NotNull BlockPos pos, BlockState newState, boolean movedByPiston) {
+        if (!state.is(newState.getBlock())) {
+            if (!level.isClientSide && level.getBlockEntity(pos) instanceof AqueductFertilizerBlockEntity blockEntity) {
+                Containers.dropContents(level, pos, blockEntity.getItemStackHandler().getItems());
+            }
+        }
+
+        super.onRemove(state, level, pos, newState, movedByPiston);
     }
 
     public static void registerModel(@NotNull BlockStateProvider provider) {