+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## []
 
+- Added Strainer for obtaining Galena/Cassiterite
+- Removed generation of Galena/Cassiterite ores in world
+- Fixed missing item dropping on block destroy
+- Added custom tag for Crafting Workbench Tool
 - Added Jade support for 1.21.3
 
 ## [0.9.2]